+++ src/test/java/kitchenpos/order/acceptanceTest/OrderAcceptanceTest.java
@@ -17,10 +17,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.acceptance.AcceptanceTest;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderLineItems;
+import kitchenpos.common.acceptance.AcceptanceTest;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
